scripts/requirements.txt
# Python dependencies for supplier data generation
PyPDF2==3.0.1
openai==1.51.0
python-dotenv==1.0.0